Refactor communication to cleanly deal with start-up and shut-down. (#101)
This PR changes some aspect of the communication module to deal with startup
and shutdown of units.

In order to allow a unit or target that is started after an upstream unit
and misses updates to gain access to the last update, a gate will store the
last update and include it in its response to the subscription request. A
consequence of this is that the last update of every unit will be kept
around. In most cases, they need to be kept around, anyway – the rtr unit
needs it for diffing, the any unit needs to keep the updates for all its
sources so it can seamlessly switch –, so this shouldn’t have too much of an
impact in practice.

The PR also fixes an issue when shutting down units. In this case, the link
goes into a “gone” state. Querying the link was immediately returning with
the gone state. If the unit owning the link was looping over this link
query, it would just end up in an endless, uninterrupted loop which would
take over one of the threads of the Tokio runtime. This is fixed by having
Link::query not return in case the status is “gone” but rather just remain
in pending state forever.

In addition, the PR does a change the the handling of suspending a link.
Earlier, querying the link would un-suspend it (or rather, that was the
intention, I believe it didn’t actually do that). Now suspending and
unsuspending is just a state change communicated to the gate which then can
in turn report this to its unit for further consideration.

Finally, the PR adds both a test unit and a test target as well as changes
the manager to allow setting up a unit graph programmatically. This way,
test cases with unit graphs can be written and their behaviour asserted.