Combine payload update and setting healthy status. (#107)
This PR revises how updates are dispatched. Instead of having separate
status and payload updates, these are now merged into one with a payload
update implying a status change to healthy. This should avoid the potential
for strange races between setting the status back to healthy and updating
the data.

A resulting limitation is that if a unit wants to change from stalled to
healthy it has to submit a payload update, but in practice this health
status change is pretty much always triggered by a new payload update
becoming available, so this should be fine.

This is a breaking change.