Allow client certificates with the json unit. (#124)
This PR adds client certificate support for the “json” unit.

It adds a new optional config variable “identity“ to the unit that contains
the path to the identity file containing at least one certificate and a
private key in PEM format. This identity will used when communicating with
an HTTPS server to acquire the JSON data.